incr super i64

+++ src/store/mongo_extends.rs
@@ -100,7 +100,7 @@ pub trait MongoStorageOpExtends<F: Filter>: Sync + Send + Clone + 'static {
         Self: 'a,
         T: MongoDbModel;
 
-    fn incr<'r, T>(self, kv_pairs: &'r [(&str, u32)], q: Condition<F>) -> Self::IncrFuture<'r, T>
+    fn incr<'r, T>(self, kv_pairs: &'r [(&str, i64)], q: Condition<F>) -> Self::IncrFuture<'r, T>
     where
         T: MongoDbModel;
